Refactor to integer delegate output.

+++ MyoTrainer/Binding/OCMyo.h
@@ -74,7 +74,7 @@ typedef enum MyoDirection {
 - (void)myoOnDisconnect:(Myo *)myo timestamp:(uint64_t)timestamp;
 - (void)myo:(Myo *)myo onRssi:(int8_t)rssi timestamp:(uint64_t)timestamp;
 - (void)myo:(Myo *)myo onPose:(MyoPose *)pose timestamp:(uint64_t)timestamp;
-- (void)myo:(Myo *)myo onEmgData:(int8_t *)emgData timestamp:(uint64_t)timestamp;
+- (void)myo:(Myo *)myo onEmgData:(int *)emgData timestamp:(uint64_t)timestamp;
 - (void)myoOnPair:(Myo *)myo firmwareVersion:(NSString *)firmware timestamp:(uint64_t)timestamp;
 - (void)myoOnConnect:(Myo *)myo firmwareVersion:(NSString *)firmware timestamp:(uint64_t)timestamp;
 - (void)myo:(Myo *)myo onGyroscopeDataWithVector:(MyoVector *)vector timestamp:(uint64_t)timestamp;
